requirements.txt
# Core dependencies
python-binance==1.0.19
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.3
pyyaml==6.0.1

# Technical indicators
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Data visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Logging and monitoring
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Data storage
jsonschema==4.20.0
